pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "evidenceai"
version = "0.1.0"
authors = [
    { name = "EvidenceAI Team" },
]
description = "PDF processing and text extraction tools"
requires-python = ">=3.8"
dependencies = [
    "PyPDF2>=3.0.0",
    "pdfminer.six>=20221105",
    "python-magic>=0.4.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
